feat(phone): phase 1 - extract shared ExpandableCommitTile widget
Extract duplicated _ExpandableCommitTile from BranchDetailScreen and
CommitHistoryScreen into shared widgets/expandable_commit_tile.dart.
Eliminates ~260 lines of duplication. Both screens now use the shared
widget with shrinkWrapDiff=true for BranchDetailScreen to maintain
nested scroll behavior.
